refactor: move missing from stub files
